+++ lmcache/integration/sglang/sglang_adapter.py
@@ -10,7 +10,6 @@ import torch
 import torch.distributed as dist
 
 # First Party
-from lmcache.config import LMCacheMetadata
 from lmcache.integration.sglang.utils import ENGINE_NAME, lmcache_get_config
 from lmcache.logging import init_logger
 from lmcache.utils import mock_up_broadcast_fn, mock_up_broadcast_object_fn
@@ -21,6 +20,7 @@ from lmcache.v1.gpu_connector import (
     SGLangGPUConnector,
     SGLangLayerwiseGPUConnector,
 )
+from lmcache.v1.metadata import LMCacheMetadata
 
 logger = init_logger(__name__)
