+++ lmcache/usage_context.py
@@ -18,10 +18,10 @@ import requests
 import torch
 
 # First Party
-from lmcache.config import LMCacheMetadata
 from lmcache.connections import global_http_connection
 from lmcache.logging import init_logger
 from lmcache.v1.config import LMCacheEngineConfig
+from lmcache.v1.metadata import LMCacheMetadata
 
 if TYPE_CHECKING:
     # First Party
