+++ lmcache/v1/lookup_client/lmcache_async_lookup_client.py
@@ -10,7 +10,6 @@ import torch
 import zmq
 
 # First Party
-from lmcache.config import LMCacheMetadata
 from lmcache.logging import init_logger
 from lmcache.v1.cache_engine import LMCacheEngine
 from lmcache.v1.config import LMCacheEngineConfig
@@ -20,6 +19,7 @@ from lmcache.v1.lookup_client.async_lookup_message import (
     LookupRequestMsg,
     LookupResponseMsg,
 )
+from lmcache.v1.metadata import LMCacheMetadata
 from lmcache.v1.rpc_utils import (
     get_zmq_context,
     get_zmq_rpc_path_lmcache,
