+++ lmcache/v1/lookup_client/mooncake_lookup_client.py
@@ -6,11 +6,11 @@ from typing import Optional, Union
 import torch
 
 # First Party
-from lmcache.config import LMCacheMetadata
 from lmcache.logging import init_logger
 from lmcache.utils import CacheEngineKey
 from lmcache.v1.config import LMCacheEngineConfig
 from lmcache.v1.lookup_client.abstract_client import LookupClientInterface
+from lmcache.v1.metadata import LMCacheMetadata
 
 logger = init_logger(__name__)
