+++ tests/v1/storage_backend/test_p2p_backend_with_controller.py
@@ -15,7 +15,6 @@ import torch
 import zmq
 
 # First Party
-from lmcache.config import LMCacheMetadata
 from lmcache.logging import init_logger
 from lmcache.utils import CacheEngineKey
 from lmcache.v1.cache_controller.message import (
@@ -27,6 +26,7 @@ from lmcache.v1.memory_management import (
     MemoryFormat,
     PagedCpuGpuMemoryAllocator,
 )
+from lmcache.v1.metadata import LMCacheMetadata
 from lmcache.v1.storage_backend.local_cpu_backend import LocalCPUBackend
 from lmcache.v1.storage_backend.p2p_backend import P2PBackend
 from lmcache.v1.transfer_channel.transfer_utils import P2PInitSideRetMsg
