+++ tests/v1/test_freeze_mode_integration.py
@@ -8,11 +8,11 @@ import requests
 import torch
 
 # First Party
-from lmcache.config import LMCacheMetadata
 from lmcache.utils import mock_up_broadcast_fn, mock_up_broadcast_object_fn
 from lmcache.v1.cache_engine import LMCacheEngineBuilder
 from lmcache.v1.config import LMCacheEngineConfig
 from lmcache.v1.internal_api_server.api_server import InternalAPIServer
+from lmcache.v1.metadata import LMCacheMetadata
 from lmcache.v1.mock_gpu_connector import MockGPUConnector
 from tests.v1.utils import (
     MockAdapter,
